requirements.txt
Django>=2.0
django-jutil>=1.0.1
django-jacc>=1.0.0

